11- Quiz
11- Generate a Package with two public classes:  Clinic class and Doctor class.
Doctor class  should contain  a string for name, that must be constant (final),  a string for working_day
( ex: Saturday, Sunday ..etc.) and a static counter to count number of generated Doctors objects. 
Implement a parametrized constructor that sets the name and working day.
Clinic class should contain (1) a private array of 3 doctors . (2) A function  Insert_doctor , that accepts a doctor object and
adds it in the array , and also  prints the value of the counter after each insertion.(3) A function display_doctors that 
prints all the doctors names that work on a given day.
Create a class with the main function in another package.
In main generate a Clinic object  and scan data of  the 3 doctors and display all doctors who works on saturday.